Allow user to say "No" to final prompt
This allows the user to specify "No" at the final prompt
when tagging the files in the current directory.  When this
happens, all the values are cleared the current process
of tagging the files in the current directory is started over.

We now have a fairly idiomatic way of choosing what to do
on user interaction:

* ALBUM METADATA user prompt:
  - Yes: Save changed tags
  - No:  Reset tag values and start the
         process of changing tags over

* TRACK METADATA user prompt:
  - Yes: Save changed tags
  - No:  Reset tag values and start the
         process of changing tags over

* FINAL PROMPT
  - Yes: Go through and tag all the FLAC files
         in the current directory
  - No:  Reset all tag values and start the
         process of changing tags over for the
         entire directory/album

Program/Script version available via the '-v' or
'--version' argument

Code cleanup/bug fixes