Lots of cleanups and optimizations!
Separated the changing of individual track tags by putting
said code into a separate function.  This negates the need
of using an associative array.

Finding the album tag values is much more exhaustive.  Now,
we search each file in the current directory, looking for a
tag value, stopping as soon as we find one.

TRACKTOTAL tag is now dynamically obtained via the number
of FLAC files found in the currently processed directory.
This allows the user to see how many files are in the directory
when deciding the TRACKTOTAL (helpful when the value is <empty>
and you want to use the # of FLAC files in the current directory).

Greatly sped up the re-tagging of FLAC files by only calling
metaflac once, making using of metaflac's ability to do multiple
operations on multiple FLAC files in one invocation.

Individual track tagging (eg. TRACKNUMBER and TITLE) are still
applied individually per file, but the operation still is faster
with one metaflac call per file instead of two.

Code cleanup/bug fixes/UI changes