Implemented multi-disc tagging option
By typing "m", "multi", or "multiple" in the DISCNUMBER
field, a new screen shows each file to tag with it's corresponding
DISCNUMBER value.  You can now choose which files are part of which
disc.

A new multi-disc prompt appears if changing the DISCNUMBER value
individually.

Bug fixes and small UI changes as well:
 - If an album has a bit depth and/or sample rate higher than a
   CD (16bit/44.1kHz), then a small message is displayed after
   the current ALBUM tag value, indicating the release may sourced
   from something other than a CD (eg. Vinyl, DVD, etc)
 - Each banner now has it's own banner message that's isolated from
   the drawing portion of the script
 - Other small improvements